Changed int variable for vacuum timer reading to double variable, because the timer run over the range of an integer. Vacuum cleaner doesn't need to run, during purge4ever process.

+++ rx_main_ctrl/step_lb.c
@@ -46,7 +46,7 @@ static int                  _WashStarted;
 static UINT32			    _Flushed = 0x00;		// For capping function which is same than flushing (need to purge after cap)
 static int                  _ScrewCommandSend[STEPPER_CNT] = {FALSE};
 static int                  _OldVacuum_Cleaner_State = FALSE;
-static INT32                _Vacuum_Cleaner_Time = 0;
+static double                _Vacuum_Cleaner_Time = 0;
 static int                  _AutoCapMode;
 
 static int                  _AutoCapTimer = 0;
@@ -270,7 +270,7 @@ int steplb_handle_status(int no, SStepperStat *pStatus)
         }
         else if (_OldVacuum_Cleaner_State == TRUE && !_Status[i].info.vacuum_running && i == 1)
         {
-            INT32 _OldVacuum_Cleaner_Time = 0;
+            double _OldVacuum_Cleaner_Time = 0;
             setup_vacuum_cleaner(PATH_USER FILENAME_VACUUM_TIME, &_OldVacuum_Cleaner_Time, READ);
             _OldVacuum_Cleaner_Time = _OldVacuum_Cleaner_Time + rx_get_ticks()/1000 - _Vacuum_Cleaner_Time;
             setup_vacuum_cleaner(PATH_USER FILENAME_VACUUM_TIME, &_OldVacuum_Cleaner_Time, WRITE);
